Got data group saving and loading in text format
Data groups can now save and load in text format.
Also got the binary version writing.
Fixed a lot of bugs with getting and setting properties in data groups - thinking maybe adding arrays to it.
Added loading files in File Utilities.
Data group Property now attempts to returns the value you want even if ti wasn't the value it was saved as - not sure if ti was worth doing or not. Maybe we should just assert and return.
Added a test cast for loading and saving