Datagroups have inheritance!
Just as it says data groups now have inheritance. We do it two ways. You can specify it by file name or by database name. Had to hack something in for the database but I think it'll be safe.

also had to do a parent cache for datagroups to be able to save and not include parent values.